+++ packages/google-pay-integration/src/google-pay-paypal-commerce/types.ts
@@ -1,0 +1,61 @@
+export interface PayPalCommerceHostWindow extends Window {
+    paypal?: PayPalGoogleSdk;
+}
+
+export interface GooglePayConfig {
+    allowedPaymentMethods: AllowedPaymentMethods[];
+    apiVersion: number;
+    apiVersionMinor: number;
+    countryCode: string;
+    isEligible: boolean;
+    merchantInfo: {
+        merchantId: string;
+        merchantOrigin: string;
+    };
+}
+
+export interface AllowedPaymentMethods {
+    type: string;
+    parameters: {
+        allowedAuthMethods: string[];
+        allowedCardNetworks: string[];
+        billingAddressRequired: boolean;
+        assuranceDetailsRequired: boolean;
+        billingAddressParameters: {
+            format: string;
+        };
+    };
+    tokenizationSpecification: {
+        type: string;
+        parameters: {
+            gateway: string;
+            gatewayMerchantId: string;
+        };
+    };
+}
+
+export interface PayPalGoogleSdk {
+    version: string;
+    getCorrelationID: () => void;
+    Googlepay: () => {
+        config: () => Promise<GooglePayConfig>;
+        confirmOrder: (arg0: {
+            orderId: string;
+            paymentMethodData: ConfirmOrderData;
+        }) => Promise<{ status: string }>;
+        initiatePayerAction: () => void;
+    };
+    FUNDING: Record<string, string>;
+}
+
+export interface ConfirmOrderData {
+    tokenizationData: {
+        type: string;
+        token: string;
+    };
+    info: {
+        cardNetwork: string;
+        cardDetails: string;
+    };
+    type: string;
+}