+++ packages/paypal-commerce-integration/src/paypal-commerce-alternative-methods/paypal-commerce-alternative-methods-payment-initialize-options.ts
@@ -1,4 +1,4 @@
-import { BirthDate, PayPalCommerceFieldsStyleOptions } from '../paypal-commerce-types';
+import { PayPalCommerceFieldsStyleOptions } from '../paypal-commerce-types';
 
 /**
  * A set of options that are required to initialize the PayPal Commerce payment
@@ -116,45 +116,7 @@ export default interface PayPalCommerceAlternativeMethodsPaymentOptions {
     submitForm(): void;
 }
 
-export interface PaypalCommerceRatePay {
-    /**
-     * The CSS selector of a container where the payment widget should be inserted into.
-     */
-    container: string;
-
-    /**
-     * The CSS selector of a container where the legal text should be inserted into.
-     */
-    legalTextContainer: string;
-
-    /**
-     * Callback that triggers pending logic (for example: show loading overlay)
-     */
-    onPaymentSubmission?(isPaymentSubmitting: boolean): void;
-
-    /**
-     * A callback that gets form values
-     */
-    getFieldsValues?(): {
-        ratepayBirthDate: BirthDate;
-        ratepayPhoneNumber: string;
-        ratepayPhoneCountryCode: string;
-    };
-
-    /**
-     * A callback right before render Smart Payment Button that gets called when
-     * Smart Payment Button is eligible. This callback can be used to hide the standard submit button.
-     */
-    onRenderButton?(): void;
-
-    /**
-     * A callback for displaying error popup. This callback requires error object as parameter.
-     */
-    onError?(error: unknown): void;
-}
-
 export interface WithPayPalCommerceAlternativeMethodsPaymentInitializeOptions {
     paypalcommerce?: PayPalCommerceAlternativeMethodsPaymentOptions; // FIXME: this option is deprecated
     paypalcommercealternativemethods?: PayPalCommerceAlternativeMethodsPaymentOptions;
-    paypalcommerceratepay?: PaypalCommerceRatePay;
 }